requirements.txt
Django>=3.2
djangorestframework>=3.12
django-filter>=2.4
drf-yasg>=1.20.0
django-cors-headers>=3.5.0
